requirements.txt
azure-identity
azure-servicebus
dapper